Local Feed Loader does not message store upon creation before loading the feed from the cache store.
