+++ EssentialDeveloper/EssentialDeveloperTests/Feed Cache/LoadFeedFromCacheTests.swift
@@ -1,0 +1,23 @@
+import XCTest
+import EssentialDeveloper
+
+final class LoadFeedFromCacheTests: XCTestCase {
+
+    func test_init_doesNotMessageTheStoreUponCreation() {
+        let (_, store) = makeSUT()
+        
+        XCTAssertEqual(store.receivedMessages, [])
+    }
+    private func makeSUT(
+        currentDate: @escaping () -> Date = Date.init,
+        file: StaticString = #file,
+        line: UInt = #line
+    ) -> (sut: LocalFeedLoader, store: FeedStoreSpy) {
+        let store = FeedStoreSpy()
+        let sut = LocalFeedLoader(store: store, currentDate: currentDate)
+        trackForMemoryLeaks(store, file: file, line: line)
+        trackForMemoryLeaks(sut, file: file, line: line)
+        return (sut, store)
+    }
+    
+}
